class AdminAuthenticator et securite sur routes admin

+++ controller/AdminController.php
@@ -1,5 +1,6 @@
 <?php
 
+require_once './service/AdminAuthenticator.php';
 require_once './repository/ProductRepository.php';
 require_once './repository/UserRepository.php';
 require_once './model/User.php';
@@ -13,6 +14,7 @@ class AdminController extends AbstractController
     public function __construct()
     {
         $this->repository = new UserRepository();
+        $this->role = new Admin();
     }
     
     
@@ -21,17 +23,15 @@ class AdminController extends AbstractController
      */
     public function displayAdmin() 
     {
-        $users = $this->repository->fetchAll();
-        
-                
-        $user = unserialize($_SESSION['user']);
-        if ($user->getRole() !== 'admin') {
-            $this->displayTwig('login');
-        } else {
+        if ($this->role->isAdmin()) {
+            $users = $this->repository->fetchAll();
+            
             $this->displayTwig('admin/adminUsers', [
-            'session' => unserialize($_SESSION['user']),
-            'users' => $users]);
-        }
+                'session' => unserialize($_SESSION['user']),
+                'users' => $users]);
+        } else {
+            $this->displayTwig('home');
+        }        
     }
     
     
@@ -40,13 +40,12 @@ class AdminController extends AbstractController
      */
     public function displayAdminCategories() 
     {
-        if (!isset($_SESSION['user'])) {
-            $this->displayTwig('login');
-            
-        } else {
+        if ($this->role->isAdmin()) {
             $this->displayTwig('admin/adminCategories', [
-            'session' => unserialize($_SESSION['user'])]);
-        }
+                'session' => unserialize($_SESSION['user'])]);
+        } else {
+            $this->displayTwig('home');
+        }        
     }
     
     
@@ -55,17 +54,16 @@ class AdminController extends AbstractController
      */
     public function displayAdminProducts() 
     {
-        $productRepository = new ProductRepository();
-        $products = $productRepository->fetchAll();
-
-        if (!isset($_SESSION['user'])) {
-            $this->displayTwig('login');
+        if ($this->role->isAdmin()) {
+            $productRepository = new ProductRepository();
+            $products = $productRepository->fetchAll();
             
-        } else {
             $this->displayTwig('admin/adminProducts', [
-            'session' => unserialize($_SESSION['user']),
-            'products' => $products]);
-        }
+                'session' => unserialize($_SESSION['user']),
+                'products' => $products]);
+        } else {
+            $this->displayTwig('home');
+        }        
     }
     
     
@@ -74,13 +72,16 @@ class AdminController extends AbstractController
      */
     public function displayUpdateUser(): void
     {
-        $_SESSION['csrf'] = bin2hex(random_bytes(32));
-        
-        $user = $this->repository->fetchById($_GET['id']);
-        
-        $this->displayTwig('admin/updateAccountAdmin', [
-            'user' => $user,
-            'csrf' => $_SESSION['csrf']]);
+        if ($this->role->isAdmin()) {
+            $_SESSION['csrf'] = bin2hex(random_bytes(32));
+            $user = $this->repository->fetchById($_GET['id']);
+            
+            $this->displayTwig('admin/updateAccountAdmin', [
+                'user' => $user,
+                'csrf' => $_SESSION['csrf']]);
+        } else {
+            $this->displayTwig('home');
+        }        
     }
     
     
@@ -89,24 +90,28 @@ class AdminController extends AbstractController
      */
     public function updateAccount()
     {
-        
-        if(!$_SESSION['csrf'] || $_SESSION['csrf'] !== $_POST['csrf_token']){
+        if ($this->role->isAdmin()) {
+            if(!$_SESSION['csrf'] || $_SESSION['csrf'] !== $_POST['csrf_token']){
             header('location: ./index.php?url=admin/account/update');
             exit();
-        }
-        
-        if($_POST['firstName'] || $_POST['lastName'] || $_POST['email']) {
+            }
             
-            $user = new User();
-            $user->setId($_GET['id']);
-            $user->setLastName(htmlspecialchars($_POST['lastName']));
-            $user->setFirstName(htmlspecialchars($_POST['firstName']));
-            $user->setEmail(htmlspecialchars($_POST['email']));
+            if($_POST['firstName'] || $_POST['lastName'] || $_POST['email']) {
+                
+                $user = new User();
+                $user->setId($_GET['id']);
+                $user->setLastName(htmlspecialchars($_POST['lastName']));
+                $user->setFirstName(htmlspecialchars($_POST['firstName']));
+                $user->setEmail(htmlspecialchars($_POST['email']));
+                
+                $this->repository->updateProfil($user);
+                
+                header('location: ./index.php?url=admin/users');
+                exit();
+            } 
             
-            $this->repository->updateProfil($user);
-            
-            header('location: ./index.php?url=admin/users');
-            exit();
+        } else {
+            $this->displayTwig('home');
         }
     }
     
@@ -115,13 +120,19 @@ class AdminController extends AbstractController
      * @Route ("index.php?url=admin/account/delete")
      */
     public function deleteAccount()
-    {
-        $user = new User();
-        $user->setId($_GET['id']);
-        
-        $data = $this->repository->deleteAccount($user);
+    { 
+        if ($this->role->isAdmin()) {
+            $user = new User();
+            $user->setId($_GET['id']);
             
-        header('location: ./index.php?url=admin/users');
-        exit();
+            $data = $this->repository->deleteAccount($user);
+                
+            header('location: ./index.php?url=admin/users');
+            exit();
+            
+        } else {
+            $this->displayTwig('home');
+        }        
     }
+    
 }
